ClientTmcSettings moved
