+++ src/test/java/hy/tmc/core/commands/GetCourseTest.java
@@ -13,7 +13,7 @@ import hy.tmc.core.TmcCore;
 import hy.tmc.core.domain.Course;
 import hy.tmc.core.domain.Exercise;
 import hy.tmc.core.exceptions.TmcCoreException;
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import java.util.ArrayList;
 import java.util.List;
 import static org.junit.Assert.assertEquals;
