+++ src/test/java/hy/tmc/core/commands/ListExercisesTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.fail;
 import hy.tmc.core.communication.ExerciseLister;
 import hy.tmc.core.communication.TmcJsonParser;
 
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import hy.tmc.core.domain.Course;
 import hy.tmc.core.exceptions.TmcCoreException;
