+++ src/test/java/hy/tmc/core/commands/RunTestsTest.java
@@ -1,6 +1,6 @@
 package hy.tmc.core.commands;
 
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 
 import hy.tmc.core.exceptions.TmcCoreException;
