+++ src/test/java/hy/tmc/core/commands/SendSpywareDiffsTest.java
@@ -12,7 +12,7 @@ import hy.tmc.core.communication.HttpResult;
 import hy.tmc.core.communication.TmcJsonParser;
 import hy.tmc.core.domain.Course;
 import hy.tmc.core.exceptions.TmcCoreException;
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import hy.tmc.core.testhelpers.ExampleJson;
 import java.util.ArrayList;
 import java.util.List;
