+++ src/test/java/hy/tmc/core/communication/ExerciseListerTest.java
@@ -1,7 +1,7 @@
 package hy.tmc.core.communication;
 
 import com.google.common.base.Optional;
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import hy.tmc.core.domain.Course;
 import hy.tmc.core.domain.Exercise;
 import hy.tmc.core.exceptions.TmcCoreException;
