+++ src/test/java/hy/tmc/core/communication/UrlCommunicatorTest.java
@@ -11,7 +11,7 @@ import static org.junit.Assert.assertEquals;
 
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.google.common.base.Optional;
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import hy.tmc.core.exceptions.TmcCoreException;
 import java.io.File;
 import java.io.IOException;
