+++ src/test/java/hy/tmc/core/zipping/ProjectRootFinderTest.java
@@ -2,7 +2,7 @@ package hy.tmc.core.zipping;
 
 import com.google.common.base.Optional;
 import hy.tmc.core.communication.TmcJsonParser;
-import hy.tmc.core.ClientTmcSettings;
+import hy.tmc.core.testhelpers.ClientTmcSettings;
 import hy.tmc.core.domain.Course;
 import hy.tmc.core.exceptions.TmcCoreException;
 import java.io.IOException;
